Move time viewer code into Brain
