+++ mne/viz/_brain/__init__.py
@@ -11,7 +11,6 @@
 
 from ._brain import Brain
 from ._scraper import _BrainScraper
-from ._timeviewer import _TimeViewer
 from ._linkviewer import _LinkViewer
 
 __all__ = ['Brain']
